+++ web/mirage/factories/me.ts
@@ -3,7 +3,7 @@ import {
   TEST_USER_EMAIL,
   TEST_USER_NAME,
   TEST_USER_GIVEN_NAME,
-} from "../mirage-utils";
+} from "../utils";
 
 export default Factory.extend({
   id: TEST_USER_EMAIL,
