+++ web/tests/acceptance/authenticated/new/project-test.ts
@@ -3,10 +3,7 @@ import { click, currentURL, fillIn, visit, waitFor } from "@ember/test-helpers";
 import { MirageTestContext, setupMirage } from "ember-cli-mirage/test-support";
 import { setupApplicationTest } from "ember-qunit";
 import { authenticateSession } from "ember-simple-auth/test-support";
-import {
-  TEST_WEB_CONFIG,
-  authenticateTestUser,
-} from "hermes/mirage/mirage-utils";
+import { TEST_WEB_CONFIG, authenticateTestUser } from "hermes/mirage/utils";
 import { Response } from "miragejs";
 import { module, test } from "qunit";
