+++ web/tests/integration/components/editable-field-test.ts
@@ -2,10 +2,7 @@ import { click, fillIn, render, triggerKeyEvent } from "@ember/test-helpers";
 import { hbs } from "ember-cli-htmlbars";
 import { MirageTestContext, setupMirage } from "ember-cli-mirage/test-support";
 import { setupRenderingTest } from "ember-qunit";
-import {
-  TEST_USER_EMAIL,
-  authenticateTestUser,
-} from "hermes/mirage/mirage-utils";
+import { TEST_USER_EMAIL, authenticateTestUser } from "hermes/mirage/utils";
 import { module, test } from "qunit";
 
 const EDITABLE_FIELD = ".editable-field";
