+++ web/tsconfig.json
@@ -30,6 +30,6 @@
     "node_modules/@gavant/glint-template-types/types/ember-render-modifiers/*",
     "node_modules/@gavant/glint-template-types/types/ember-truth-helpers/*",
     "node_modules/@gavant/glint-template-types/types/ember-power-select/**/*",
-    "mirage/mirage-utils.ts"
+    "mirage/utils.ts"
   ]
 }
